Updated 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	unsigned long fibonacci_0, fibonacci_1, fibonacci_2;
+	float fibonacci_0, fibonacci_1, fibonacci_2;
 
 	fibonacci_0 = 1;
 	fibonacci_1 = 2;
@@ -17,9 +17,9 @@ int main(void)
 	{
 		fibonacci_2 = fibonacci_0 + fibonacci_1;
 		if (fibonacci_0 == 1)
-			printf("%lu", fibonacci_0);
+			printf("%.0f", fibonacci_0);
 		else
-			printf(", %lu", fibonacci_0);
+			printf(", %.0f", fibonacci_0);
 		fibonacci_0 = fibonacci_1;
 		fibonacci_1 = fibonacci_2;
 	}
